refactor: migrate EnumAsset system to 64-bit long IDs
- Switched InternalValue from int to long to support 1B multiplier logic.
- Updated EnumAssetEditor to handle long masking and Asset ID prefixing.
- Refactored EnumPicker, EnumTable, and EnumToEnumMap property drawers
  to use SerializedProperty.longValue, preventing integer overflow.
- Improved EnumTable deletion logic to safely handle object references.
- Resolved NullReferenceException in EnumTableDrawer by using relative
  property pathing for asset source lookup.

+++ Assets/_Project/Scripts/Features/Actors/New/State/EntityIdleState.cs
@@ -9,7 +9,7 @@ namespace Kope.Actor.New {
 
 		public override void TickUpdate() {
 			if (this._movementComponent.Direction.sqrMagnitude >= MovementComponentBase.MOVEMENT_EPSILON) {
-				int movementEnumId = this.movementStateEnum.GetSelectedEnumId();
+				long movementEnumId = this.movementStateEnum.GetSelectedEnumId();
 
 				// already in Idle; we don't need to 'fallback' to it if Move is briefly Busy.
 				_ = this._stateManagement.ChangeState(movementEnumId, false);
